pyproject.toml
[tool.poetry]
name = "whats-for-dinner"
version = "0.1.0"
description = ""
authors = ["Shipra Shalini <mailx.shipras@gmail.com>"]
readme = "README.md"
packages = [{ include = "whats_for_dinner", from = "src" }]


[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "2.5.1"
fastapi = "^0.115.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
pgvector-haystack = "^1.0.0"
pgvector = "^0.3.4"
uvicorn = "^0.31.0"


[tool.black]
color = true
line-length = 88
target-version = ['py312']


[tool.isort]
line_length = 88
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
